requirements.txt
anthropic>=0.39.0
streamlit>=1.30.0
pandas>=2.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
rapidfuzz>=3.5.0
jinja2>=3.1.0
click>=8.1.0
pytest>=7.4.0
